Random Turret Looking
Now, rather than just spinning around, Turrets look around randomly.
I think this is more interesting than just sweeping, but more
importantly I originally felt that there was strategy in the sweep where
it misses the bad guy and you're like "oh no!"

But in practice, no. It's just frustrating and feels like the game
screwed you.

In order to do that I built this MomentumTween object that sped up and
slowed down, which was fun