Switch OmniVision
Through playtesting it became clear that there was a pretty major flaw
in the way I wanted switches to play out. There would be a switch
attached to a shield and a turret, and the enemy would knock out the
shield and then move in to attack the turret and... the whole thing
would just shut down...

The turret would stop shooting and the shield wouldn't recharge.
And the reason is that the enemy had walked past the switch, and now
everything was off, which made sense but also really sucked.

So it was clear that sensors needed omnivision rather than just the
wedge that I wanted enemies to have. So I copied nearly 100% of the
vision code into an Omnivision thing, and now we're all good.